+++ doc/ChangeLog.md
@@ -2,6 +2,11 @@
 
 Start for 3.2 version. No change log before.
 
+### 3.3.3 :   
+- add new install mode : default, server, custom
+- add gimp, darktable, ser player, Stackistry, Imppg
+- add Astrophoto-plus with PHD2 support
+
 ### 3.2.2 :   
 - Add ddserver for qdslrdashboard (in web communications)
 - Add package for ExFat